Suppress `ctor_ident` in docs (Function)
The code for ctors on statics already uses the `#[doc(hidden)]`
attribute to suppress symbols created by ctor out of rustdoc
output, but the one for functions does not.

If you add that annotation to the function wrapped by `ctor`,
it'll suppress the function at global scope, but not the static
that ctor generates.

This patch adds `#[doc(hdiden)]` to the latter as well.